MED_AI/requirements.txt
streamlit>=1.35.0
langgraph>=0.1.14
langchain>=0.2.0
langchain-openai>=0.1.14
langchain-anthropic>=1.4.0
langchain-community>=0.2.0
tavily-python>=0.3.3
openai>=1.30.0
pdfplumber>=0.11.0
PyMuPDF>=1.24.0          
Pillow>=10.0.0
pytesseract>=0.3.10      